Implement CachedDiagnostics
Implement CachedDiagnostics that can be used to store and replay the
diagnostics without compiling the code. When a diagnostics is emitted,
the CachedDiagnosticConsumer can capture and serialize `DiagnosticInfo`
to YAML format, with an option to instant replay the diagnostics to
other Diagnostics consumer from the cache with a special SourceManager.

CachedDiagnostics allows caching build to store diagnostics without
worring about the diagnostics options like "color diagnostics" and it
can also be used to remap diagnostics to a different file path.